Modify the logic for STAT-Analysis to enable the use of command line options in conjunction with a config file.  The logic goes as follows:  - Set up a "default_job" to be used when filtering through the input STAT lines.  - If a config file is specified, parse the default job from the top filtering section of the config file.  - Whether or not a config file was specified, amend the default job with any job command settings left over from the command line.  - Apply that filter job to the input stat lines and write the remaining lines to a temp file.  - Read data from that temp file to process all other jobs (from the config file or command line).  - Initialize each job using the "default_job".  - Then used the jobstring passed in to the process_job() function to set the job up.  - Lastly, use the command line job to amend the current job with any job command settings from the command line.
So users can define a single job on the command line or they can define multiple jobs in a config file.  When using a config file, users can specify command line options to override or amend any of those jobs.  This is useful if you want to run the same set of 10 jobs for each of 12 different lead times, for example.

Also, added detailed debugging info for this logic using Debug(4).