Fix the following problems when computing the GO Index via STAT-Analysis: (1) Prior to reading the go_index config file, we were failing to read the config file constants. (2) The go_index is defined over 48 terms, but the code was using 96 (48 x 2). (3) The go_index job only worked on the command line - not through the config file.
All of these issues were resolved with the following changes:
(1) Add in calls to parse the config file constants prior to parsing the go_index config file.
(2) Modify STATAnalysisJob::get_jobstring() to only include "-job type" when the job type is a valid one.
(3) Remove the reading of the go_Index config file from do_job_go_index() - that was causing the 96 members instead of 48.
(4) Add section of code for handling the go_index into process_job().  Putting it there enables it to work on the command line and through the config file.