Update ShapeData::convex_hull() routine in both the trunk and the METv4.1 bugfix branch.  The update is just adding the following line:   if ( alpha == 0 && beta == 0 )  continue;
This change is a result of all the MODE runs failing on zeus (METv4.1 built with intel compilers intel/12-12.0.4.191 or intel/13.0.079) with the following error:
  ERROR  : convex_hull(Polyline &) -> can't find next hull point 

Randy and I found that adding any print statement or conditional on a new line inside that loop resulted in this error going away.  So it must have something to do with the machine code that's being created.  We've found that alpha and beta never actually are 0, but by adding this line, the error goes away.

We don't really understand why this fixes the problem on zeus, but it does and should have no impact when running on other platforms. 



Former-commit-id: ae79181176319eda8da6a66303158afa5e0bebf1