Committing a GRIB2 change to both the met-6.0 bugfix branch and the trunk.  The Air Force produces GRIB2 files with some pretty funny level values.  This came up via met_help https://rt.rap.ucar.edu/rt/Ticket/Display.html?id=77365 and also resurfaced during the Air Force GALWEM evaluation.  Some of their GRIB2 records are defined on special level types as described in this table: http://www.nco.ncep.noaa.gov/pmb/docs/grib2/grib2_table4-5.shtml
For those records, the set the level type correctly but do not set the actual level value.  On my system, that results in a level value of -2147483647.  We had already committed a fix to manually set the level value to 0 when the level type = 1 (Ground or Water Surface).  This change expands that logic to cover level types 1 through 10.  The AF GALWEM project evaluated the tropopause and max wind level and so this change was required. 



Former-commit-id: 9cf3cb6ca0997069228d758074b6a835a97a9a43