Committing a bugfix to the logic in data2d_grib2.cc. OLD: rec->LvlTyp       = (8 == gfld->ipdtnum || 12 == gfld->ipdtnum ? 8 : gfld->ipdtmpl[9]); NEW: rec->LvlTyp       = gfld->ipdtmpl[9]; i This issue arose when Tina Kalb was trying to use MXUPHL from a GRIB2 file: 1:0:d=2017051612:MXUPHL:5000-2000 m above ground:27-28 hour max fcst: 2:2381639:d=2017051612:MXUPHL:3000-0 m above ground:27-28 hour max fcst:
Since the template number is 8, the level type is being set to a value of 8.  Looking in table 4.5 (http://www.nco.ncep.noaa.gov/pmb/docs/grib2/grib2_table4-5.shtml) that means the "Nominal Top of the Atmosphere".  And MET includes special logic to interpret level types of 1 through 10 as having a level value of 0.  Just setting the level type to gfld->ipdtmpl[9] is much simpler logic and fixes Tina's issue. 



Former-commit-id: 2af2a7b00c5956b3d71a56381e0705a2ac7fdb00