Modifying how the conifugration is handle for compiling python support.  Here's a lot of details from an email about this:
Following the paradigm we used for GRIB2, I think I fixed this with the following changes...

(1) In "configure.ac", if set, add MET_PYTHON_CC and MET_PYTHON_LD to CPPFLAGS and LDFLAGS, respectively:

+if test -n "$MET_PYTHON_CC"; then
+   CPPFLAGS="${CPPFLAGS} ${MET_PYTHON_CC}"
+fi
+
+if test -n "$MET_PYTHON_LD"; then
+   LDFLAGS="${LDFLAGS} ${MET_PYTHON_LD}"
+fi

(2) In "configure.ac", add $MET_PYTHON_LD to the end of PYTHON_LIBS:
+    PYTHON_LIBS="-lvx_data2d_python -lvx_python_utils ${MET_PYTHON_LD}"

(3) In 28 "Makefile.am" files for the MET applications, remove...
-if ENABLE_PYTHON
-   {BLAH}_CPPFLAGS += ${MET_PYTHON_CC}
-   {BLAH}_LDFLAGS  += $(MET_PYTHON_LD)
-endif
Where {BLAH} is the name of the application.

(4) Add reference to ${PYTHON_LIBS} to Makefile.am for tc_pairs and tc_stat.



Former-commit-id: 926a3cfb98c9e5c492348fa647ddf40509b90b58