Update appendixC.rst
In Roberts and Lean 2008, UFSS is defined as 0.5 +􏰆 fo/2 where fo is observed frequency  (fraction of observed points exceeding the threshold over the domain)--see Section 2d, halfway down the right column of page 83.  The critical word is observed.  

However, the User's Guide v9.0.3 it says UFSS is equivalent to the forecast event proportion--a subtle but important difference if there is any forecast frequency bias. UFSS will differ depending on if it is based on observed event proportion or forecasted event proportion. 

Now, I do believe, behind the scenes, MET based the UFSS on the observed frequency, consistent with Roberts and Lean 2008. Only the User's Guide needs to be revised.

I also revised the description of UFSS and formulation to match Roberts and Lean 2008. UFSS is simply halfway between perfect skill (unity) a random FSS (equal to observed event proportion).