Another hotfix to develop to get the Docker build step working again.
The failure from the configure script is caused by the version of the "alocal" executable.
The contents of develop are based on seneca, which has:
  aclocal (GNU automake) 1.16.1
By comparison, my laptop does NOT require that bootstrap be run for:
  aclocal (GNU automake) 1.16.5
But the Docker image has an older version which triggers the error:
  aclocal (GNU automake) 1.13.4

Any machine with aclocal 1.16.x works fine. Likely any machine with a version that differs from 1.16.x does not. Running bootstrap fixes that... but also changes the contents of all the Makefile.in files making them appear as a large set of diffs in the repo. This is not an issue inside the docker image since we generally remove the source code after compiling anyway.