Feature #2769 tc_diag_driver_v0.11.0 (#2812)
* Per #2769, update to tc_diag_driver version 0.11.0.

* Per #2769, add tmp_nc_diag_flag config option to retain temporary files.

* Per #2769, change default regrdding method from NEAREST to BILIN.

* Per #2769, add checks to make sure input files exist.

* Fix the handling of diagnostic name and units strings. Both are limited
to 7 characters but the units are enclosed in parantheses, increasing
the field width to 9. If the strings are too long, truncate them and
print warnings like this:
WARNING: OutFileInfo::write_cira_diag_vals() -> long diagnostic name "850TANGXXX" truncated to "850TANG"!
WARNING: OutFileInfo::write_cira_diag_vals() -> long diagnostic units string "(10^7C/MXXX)" truncated to "(10^7C/M)"!
Pad the first 2 columns out to widths of 7 and 9 and set the
inter-column spacing between columns 2 and 3 to 0. This could
technically introduce parsing problems when the units are 7 characters
long and the values have 5 digits, but this is the logic needed to
exactly replicate the existing output.

* Just whitespace

* Per #2769, remove the tmp_nc_diag_flag TC-Diag config file option in favor of the MET_KEEP_TEMP_FILE environment variable option that Howard is adding on his feature_2772_python_embedding_json branch. I did add documentation about that option here even though the functionality is coming from Howard's branch.

* Per #2769, remove tmp_nc_diag_flag that has been removed.

* Per #2769, remove mention of tmp_nc_diag_flag option that has been removed.

* Per #2769, upgrade to METbaseimage version 3.2 to provide SciPy in the Python environment which is needed by TC-Diag.

* Per #2769, add SciPy to the list of required Python packages.

* Per #2769, tweak AppendixF details.

* Per #2769, update release date for MET-12.0.0-beta3 to be 20240207 since we didn't get it out on the 6th.